New Commands: Cards
Added a new inner class for Cards. Includes methods for adding Aspects to the Card, via the Card_Aspects table. Also includes method for obtaining all associated Aspects in the form of Name, Count and Type (primary/secondary/other).

Added commands to the main switch block for making a card, getting and printing a card, and adding aspects to a card.